+++ src/game/index.ts
@@ -22,3 +22,4 @@ export * from './robo/offline';
 export * from './robo/achievements';
 
 export * from './robo/projects';
+export * from './projects';